+++ packages/core/src/__tests__/swift.test.ts
@@ -1,0 +1,13 @@
+import { describe } from 'vitest';
+import { componentToSwift } from '..';
+import { runTestsForTarget } from './test-generator';
+
+describe('swift', () => {
+  const generator = componentToSwift;
+
+  runTestsForTarget({
+    options: {},
+    target: 'swift',
+    generator,
+  });
+});