+++ packages/core/src/generators/mitosis/generator.ts
@@ -22,6 +22,7 @@ import { MitosisComponent } from '@/types/mitosis-component';
 import { MitosisNode, checkIsForNode, checkIsShowNode } from '@/types/mitosis-node';
 import { TranspilerGenerator } from '@/types/transpiler';
 import json5 from 'json5';
+import traverse from 'neotraverse/legacy';
 import { check, format } from 'prettier/standalone';
 
 import { blockToReact, componentToReact } from '../react';
@@ -214,7 +215,14 @@ export const blockToMitosis = (
   for (const key in json.slots) {
     const value = json.slots[key];
     str += ` ${key}={`;
-    if (value.length > 1) {
+
+    /**
+     * Pass the empty array as foo={[]} could be treated differently than not
+     * passing the prop at all.
+     */
+    if (value.length === 0) {
+      str += '[]';
+    } else if (value.length > 1) {
       str += '<>';
     }
     str += json.slots[key]
@@ -224,6 +232,17 @@ export const blockToMitosis = (
       str += '</>';
     }
     str += `}`;
+  }
+
+  for (const key in json.blocksSlots) {
+    const value = json.blocksSlots[key];
+    traverse(value).forEach(function (v) {
+      if (isMitosisNode(v)) {
+        this.update(blockToMitosis(v, toMitosisOptions, component, insideJsx));
+      }
+    });
+
+    str += `${key}={${generateBlockSlotsCode(value)}}`;
   }
 
   if (SELF_CLOSING_HTML_TAGS.has(json.name)) {
@@ -244,6 +263,29 @@ export const blockToMitosis = (
   str += `</${json.name}>`;
 
   return str;
+};
+
+const generateBlockSlotsCode = (blockSlot: any) => {
+  let code = '';
+  // generate array props (foo=[...])
+  if (Array.isArray(blockSlot)) {
+    code += `[${blockSlot.map(generateBlockSlotsCode).join(',')}]`;
+    // generate object props (foo={{ ... }})
+  } else if (typeof blockSlot === 'object' && blockSlot !== null) {
+    code += '{';
+
+    for (const key in blockSlot) {
+      if (blockSlot.hasOwnProperty(key)) {
+        code += `${key}: ${generateBlockSlotsCode(blockSlot[key])},`;
+      }
+    }
+
+    code += '}';
+  } else {
+    code += blockSlot;
+  }
+
+  return code;
 };
 
 const getRefsString = (json: MitosisComponent, refs = Array.from(getRefs(json))) => {